+++ examples/examples/manual-reqwest.rs
@@ -53,12 +53,11 @@ fn run_query() -> cynic::GraphQLResponse<FilmDirectorQuery> {
 }
 
 fn build_query() -> cynic::Operation<'static, FilmDirectorQuery> {
-    use cynic::{FragmentContext, QueryFragment};
-    cynic::Operation::query(FilmDirectorQuery::fragment(FragmentContext::new(
-        &FilmArguments {
-            id: Some("ZmlsbXM6MQ==".into()),
-        },
-    )))
+    use cynic::QueryBuilder;
+
+    FilmDirectorQuery::build(&FilmArguments {
+        id: Some("ZmlsbXM6MQ==".into()),
+    })
 }
 
 #[cfg(test)]